+++ Shared/Hanamilegal.Web.ApiCommon/Pagination/PaginationResult.cs
@@ -5,10 +5,10 @@ namespace Hanamilegal.Web.ApiCommon.Pagination;
 
 public class PaginationResult<T>
 {
-    public int PageNumber { get; init; }
-    public int PageSize { get; init; }
-    public int TotalItemsCount { get; init; }
+    public long PageNumber { get; init; }
+    public long PageSize { get; init; }
+    public long TotalItemsCount { get; init; }
     public IReadOnlyList<T> Items { get; init; } = [];
 
-    public int TotalPages => (int)Math.Ceiling((double)TotalItemsCount / PageSize);
+    public long TotalPages => (int)Math.Ceiling((double)TotalItemsCount / PageSize);
 }
